requirements.txt
flask
flask-cors
sentence-transformers
faiss-cpu
openai
python-dotenv
